refactor: split RealmModel monolithic file and decouple Readium types
