+++ YetAnotherEBookReader/Models/Realm/Persistable.swift
@@ -1,0 +1,13 @@
+//
+//  Persistable.swift
+//  YetAnotherEBookReader
+//
+
+import Foundation
+import RealmSwift
+
+public protocol Persistable {
+    associatedtype ManagedObject: RealmSwift.Object
+    init(managedObject: ManagedObject)
+    func managedObject() -> ManagedObject
+}
